refactor: Split up dataset configs, move task params to Task
