+++ src/euroeval/data_models.py
@@ -11,7 +11,6 @@ import pydantic
 import torch
 
 from .enums import Device, InferenceBackend, ModelType, TaskGroup
-from .prompt_templates import get_prompt_templates
 from .types import ScoreDict
 from .utils import get_package_version
 
@@ -130,13 +129,28 @@ class Task:
             The name of the task.
         task_group:
             The task group of the task.
+        template_dict:
+            The template dictionary for the task, from language to prompt template.
         metrics:
             The metrics used to evaluate the task.
+        num_few_shot_examples:
+            The number of examples to use when benchmarking the task using few-shot
+            evaluation. For a classification task, these will be drawn evenly from
+            each label.
+        max_generated_tokens:
+            The maximum number of tokens to generate when benchmarking the task
+            using few-shot evaluation.
+        labels:
+            The labels in the dataset.
     """
 
     name: str
     task_group: TaskGroup
+    template_dict: dict["Language", "PromptConfig"]
     metrics: list[MetricConfig]
+    num_few_shot_examples: int
+    max_generated_tokens: int
+    labels: list[str]
 
     def __hash__(self) -> int:
         """Return a hash of the task."""
@@ -352,27 +366,27 @@ class DatasetConfig:
             The mapping from label to ID.
         num_labels:
             The number of labels in the dataset.
-        prompt_template (optional):
-            The template for the prompt to use when benchmarking the dataset using
-            few-shot evaluation. Defaults to the template for the task and language.
-        max_generated_tokens (optional):
-            The maximum number of tokens to generate when benchmarking the dataset
-            using few-shot evaluation. Defaults to the template for the task and
-            language.
-        prompt_prefix (optional):
+        _prompt_prefix (optional):
             The prefix to use in the few-shot prompt. Defaults to the template for the
             task and language.
-        num_few_shot_examples (optional):
+        _prompt_template (optional):
+            The template for the prompt to use when benchmarking the dataset using
+            few-shot evaluation. Defaults to the template for the task and language.
+        _instruction_prompt (optional):
+            The prompt to use when benchmarking the dataset using instruction-based
+            evaluation. Defaults to the template for the task and language.
+        _num_few_shot_examples (optional):
             The number of examples to use when benchmarking the dataset using few-shot
             evaluation. For a classification task, these will be drawn evenly from
             each label. Defaults to the template for the task and language.
-        instruction_prompt (optional):
-            The prompt to use when benchmarking the dataset using instruction-based
-            evaluation. Defaults to the template for the task and language.
-        labels (optional):
+        _max_generated_tokens (optional):
+            The maximum number of tokens to generate when benchmarking the dataset
+            using few-shot evaluation. Defaults to the template for the task and
+            language.
+        _labels (optional):
             The labels in the dataset. Defaults to the template for the task and
             language.
-        prompt_label_mapping (optional):
+        _prompt_label_mapping (optional):
             A mapping from the labels to another phrase which is used as a substitute
             for the label in few-shot evaluation. Defaults to the template for the task
             and language.
@@ -385,14 +399,72 @@ class DatasetConfig:
     huggingface_id: str
     task: Task
     languages: list[Language]
-    prompt_template: str = field(default="")
-    max_generated_tokens: int = field(default=0)
-    prompt_prefix: str = field(default="")
-    num_few_shot_examples: int = field(default=0)
-    instruction_prompt: str = field(default="")
-    labels: list[str] = field(default_factory=list)
-    prompt_label_mapping: dict[str, str] = field(default_factory=dict)
+    _prompt_prefix: str | None = None
+    _prompt_template: str | None = None
+    _instruction_prompt: str | None = None
+    _num_few_shot_examples: int | None = None
+    _max_generated_tokens: int | None = None
+    _labels: list[str] | None = None
+    _prompt_label_mapping: dict[str, str] | None = None
     unofficial: bool = False
+
+    @property
+    def prompt_prefix(self) -> str:
+        """The prefix to use in the few-shot prompt."""
+        main_language = self.languages[0]
+        prompt_config = self.task.template_dict[main_language]
+        prompt_prefix = self._prompt_prefix or prompt_config.prompt_prefix
+        prompt_prefix = prompt_prefix.format(labels_str=self._labels_str)
+        return prompt_prefix
+
+    @property
+    def prompt_template(self) -> str:
+        """The template used during few-shot evaluation."""
+        main_language = self.languages[0]
+        prompt_config = self.task.template_dict[main_language]
+        prompt_template = self._prompt_template or prompt_config.prompt_template
+        prompt_template = prompt_template.format(labels_str=self._labels_str)
+        return prompt_template
+
+    @property
+    def instruction_prompt(self) -> str:
+        """The prompt to use when evaluating instruction-tuned models."""
+        main_language = self.languages[0]
+        prompt_config = self.task.template_dict[main_language]
+        instruction_prompt = (
+            self._instruction_prompt or prompt_config.instruction_prompt
+        )
+        instruction_prompt = instruction_prompt.format(labels_str=self._labels_str)
+        return instruction_prompt
+
+    @property
+    def num_few_shot_examples(self) -> int:
+        """The number of few-shot examples to use."""
+        return self._num_few_shot_examples or self.task.num_few_shot_examples
+
+    @property
+    def max_generated_tokens(self) -> int:
+        """The maximum number of tokens to generate when evaluating a model."""
+        return self._max_generated_tokens or self.task.max_generated_tokens
+
+    @property
+    def labels(self) -> list[str]:
+        """The labels in the dataset."""
+        return self._labels or self.task.labels
+
+    @property
+    def prompt_label_mapping(self) -> dict[str, str]:
+        """Mapping from English labels to localised labels."""
+        if self._prompt_label_mapping is not None:
+            return self._prompt_label_mapping
+
+        main_language = self.languages[0]
+        prompt_config = self.task.template_dict[main_language]
+
+        if prompt_config.prompt_label_mapping == "auto":
+            return {label: label for label in self.labels}
+        else:
+            return prompt_config.prompt_label_mapping
 
     @property
     def id2label(self) -> dict[int, str]:
@@ -413,7 +485,8 @@ class DatasetConfig:
         """Return a hash of the dataset configuration."""
         return hash(self.name)
 
-    def _get_labels_str(self, language: Language) -> str:
+    @property
+    def _labels_str(self) -> str:
         """Converts a set of labels to a natural string, in the specified language.
 
         If the task is NER, we separate using 'and' and use the mapped labels instead of
@@ -433,10 +506,12 @@ class DatasetConfig:
             >>> get_labels_str(language=DA)
             "'a', 'b', 'c' eller 'd'"
         """
+        main_language = self.languages[0]
+
         if self.task.task_group == TaskGroup.TOKEN_CLASSIFICATION:
-            sep_word = language.and_separator
+            sep_word = main_language.and_separator
         else:
-            sep_word = language.or_separator
+            sep_word = main_language.or_separator
 
         # Convert labels to single-quoted labels - and remove duplicates
         quoted_labels = [
@@ -451,39 +526,6 @@ class DatasetConfig:
             return f"{quoted_labels[0]} {sep_word} {quoted_labels[1]}"
         else:
             return f"{', '.join(quoted_labels[:-1])} {sep_word} {quoted_labels[-1]}"
-
-    def __post_init__(self) -> None:
-        """Post initialisation setup of defaults."""
-        # Skip setting defaults if the task is "speed" as it doesn't have any defaults
-        if self.task.task_group != TaskGroup.SPEED:
-            # TODO: should we just use the first language in the list?
-            main_language = self.languages[0]
-            template = get_prompt_templates(task=self.task, language=main_language)
-
-            if not self.labels:
-                self.labels = template.labels
-            if not self.prompt_label_mapping:
-                self.prompt_label_mapping = template.prompt_label_mapping
-            if not self.max_generated_tokens:
-                self.max_generated_tokens = template.max_generated_tokens
-            if not self.num_few_shot_examples:
-                self.num_few_shot_examples = template.num_few_shot_examples
-            if not self.prompt_template:
-                self.prompt_template = template.prompt_template
-            if not self.prompt_prefix:
-                self.prompt_prefix = template.prompt_prefix
-            if not self.instruction_prompt:
-                self.instruction_prompt = template.instruction_prompt
-
-            labels_str = self._get_labels_str(language=main_language)
-
-            def replace_labels(text: str) -> str:
-                """Replace '{labels_str}' in the text with the actual labels."""
-                return text.replace("{labels_str}", labels_str)
-
-            self.prompt_template = replace_labels(text=self.prompt_template)
-            self.prompt_prefix = replace_labels(text=self.prompt_prefix)
-            self.instruction_prompt = replace_labels(text=self.instruction_prompt)
 
 
 @dataclass
@@ -601,3 +643,31 @@ class HFModelInfo:
     pipeline_tag: str
     tags: list[str]
     adapter_base_model_id: str | None
+
+
+@dataclass
+class PromptConfig:
+    """Configuration for task-specific prompting across languages.
+
+    Defines the prompt templates needed for evaluating a specific task in a given
+    language.
+
+    Attributes:
+        prompt_prefix:
+            The prefix to use in the few-shot prompt.
+        prompt_template:
+            The template for the prompt to use when benchmarking the dataset using
+            few-shot evaluation.
+        instruction_prompt:
+            The prompt to use when benchmarking the dataset using instruction-based
+            evaluation.
+        prompt_label_mapping:
+            A mapping from the labels to another phrase which is used as a substitute
+            for the label in few-shot evaluation. If set to "auto", the mapping will be
+            set to a 1:1 mapping between the labels and themselves.
+    """
+
+    prompt_prefix: str
+    prompt_template: str
+    instruction_prompt: str
+    prompt_label_mapping: dict[str, str] | t.Literal["auto"]
